refactor(sub_agent): extract Definition runner to break compiler facade edge
Introduce PtcRunner.SubAgent.Runner as the internal execution boundary for
%Definition{} agents (validation, context prep, single-shot fast path, and
Loop dispatch). The public facade now delegates Definition execution to the
runner, and Compiler.compile/2 calls Runner.run/2 instead of the facade,
dropping its `alias PtcRunner.SubAgent`.

This removes the compiler->facade back-edge in the SubAgent dependency cycle
(verified: `mix xref trace compiler.ex` no longer hits sub_agent.ex) and is
the foundational change reused by #1052 for nested-agent execution.

Verified with mix precommit (format, compile --warnings-as-errors, xref
cycles, credo --strict, tests) plus demo and ptc_viewer suites.
